requirements.txt
astrbot
Pillow>=9.0.0
requests>=2.25.0